+++ src/algorithms/CcsdDiagrammaticDecomposition.cxx
@@ -345,7 +345,7 @@ void CcsdDiagrammaticDecomposition::run() {
 
   // Read the integralsSliceSize. If not provided use No
   int integralsSliceSize(getIntegerArgument("integralsSliceSize"));
-  
+
   int numberSlices(int(ceil(double(Nv)/integralsSliceSize)));
 
   Tensor<> realDressedGammaGab(realGammaGab);
@@ -375,35 +375,38 @@ void CcsdDiagrammaticDecomposition::run() {
   //slice loop starts here
   for (int m(0); m < numberSlices; m++){
     for (int n(m); n < numberSlices; n++){
-      int lenscd[] = {
-        realSlicedGammaGab[n]->lens[1], realSlicedGammaGab[m]->lens[1],
-        realSlicedGammaGab[n]->lens[2], realSlicedGammaGab[m]->lens[2]
-      };
+      int lenscd[] = {(int)realSlicedGammaGab[n]->lens[1],
+                      (int)realSlicedGammaGab[m]->lens[1],
+                      (int)realSlicedGammaGab[n]->lens[2],
+                      (int)realSlicedGammaGab[m]->lens[2]};
       int syms[] = {NS, NS, NS, NS};
       auto Vxycd(new Tensor<>(4, lenscd, syms, *realDressedGammaGab.wrld, "Vxycd"));
- 
+
       (*Vxycd)["xycd"]  = (*realSlicedGammaGab[n])["Gxc"] * (*realSlicedGammaGab[m])["Gyd"];
       (*Vxycd)["xycd"] += (*imagSlicedGammaGab[n])["Gxc"] * (*imagSlicedGammaGab[m])["Gyd"];
- 
-      int lensij[] = { Vxycd->lens[0], Vxycd->lens[1], No, No};
+
+      int lensij[] = {(int)Vxycd->lens[0],
+                      (int)Vxycd->lens[1],
+                      (int)No,
+                      (int)No};
       Tensor<> Rxyij(4, lensij, syms, *Vxycd->wrld, "Rxyij");
- 
+
       // Contract sliced Vxycd with T2 and T1 Amplitudes using Xabij
       Rxyij["xyij"] = (*Vxycd)["xycd"] * Xabij["cdij"];
       int a(n*integralsSliceSize);
       int b(m*integralsSliceSize);
- 
+
       sliceIntoResiduum(Rxyij, a, b, *Rabij);
       // The integrals of this slice are not needed anymore
       delete Vxycd;
- 
+
     }
   }
 
   Tensor<> Xklij(false, *Vijkl);
   Xklij.set_name("Xklij");
   // we have to remove the singles sinlges term which appears in our abcd contraction
-  (*Rai)["ai"] = 0.0; 
+  (*Rai)["ai"] = 0.0;
   Xklij["klij"]  = (*Vabij)["cdkl"] * Xabij["cdij"];
   (*Rabij)["abij"] += (-1.0) * Xklij["klij"] * (*Tai)["ai"] * (*Tai)["bj"];
 
@@ -440,7 +443,7 @@ void CcsdDiagrammaticDecomposition::evaluateEnergy(
 
   // get the Coulomb integrals to compute the energy
   auto Vijab(getTensorArgument<>("HHPPCoulombIntegrals"));
-  
+
   Scalar<> energy(*Vijab->wrld);
   //direct term
   energy[""]  = 2.0 * Rabij["abij"] * (*Vijab)["ijab"];
@@ -483,4 +486,3 @@ void CcsdDiagrammaticDecomposition::sliceIntoResiduum(
     Rabij.slice(dstStart,dstEnd,1.0, Ryxji,srcStart,srcEnd,1.0);
   }
 }
-
