+++ src/algorithms/FiniteSizeCorrection.cxx
@@ -128,7 +128,7 @@ void FiniteSizeCorrection::calculateRealStructureFactor() {
   int No(epsi->lens[0]);
   int Nv(epsa->lens[0]);
   int Np(No+Nv);
-  
+
   int orbitalPairStart(getIntegerArgument("orbitalPairStart",-1));
   int orbitalPairEnd(getIntegerArgument("orbitalPairEnd",-1));
 
@@ -139,14 +139,14 @@ void FiniteSizeCorrection::calculateRealStructureFactor() {
   if (orbitalPairEnd > No || orbitalPairEnd <= orbitalPairStart){
     orbitalPairEnd = No;
   }
-  
+
   int numberOrbitalPairs(orbitalPairEnd - orbitalPairStart);
   bool orbitalPairs(numberOrbitalPairs != No);
   if ( orbitalPairs ){
     LOG(0,"Orbital Pair Analysis") << "Treating only pairs " << orbitalPairStart
                                   << " to " << orbitalPairEnd << std::endl;
   }
-  
+
   PTR(Tensor<complex>) GammaGai;
   int aStart(Np-Nv), aEnd(Np);
   int iStart(0), iEnd(No);
@@ -160,7 +160,7 @@ void FiniteSizeCorrection::calculateRealStructureFactor() {
     Tensor<complex> *GammaFqr(getTensorArgument<complex>("CoulombVertex"));
     // Get the Particle Hole Coulomb Vertex
     int NF(GammaFqr->lens[0]);
-    
+
     int FaiStart[] = {0, aStart,iStart};
     int FaiEnd[]   = {NF,aEnd,  iEnd};
 
@@ -170,14 +170,14 @@ void FiniteSizeCorrection::calculateRealStructureFactor() {
       Tensor<complex> *UGF(
         getTensorArgument<complex>("CoulombVertexSingularVectors")
       );
-      int lens[]= {UGF->lens[0], Nv, No};
+      int lens[]= {(int)UGF->lens[0], (int)Nv, (int)No};
       GammaGai = NEW(Tensor<complex>,
         3, lens, GammaFqr->sym, *GammaFqr->wrld, "GammaGqr"
       );
       (*GammaGai)["Gai"] = GammaFai["Fai"] * (*UGF)["GF"];
     }
     else {
-      int lens[]= {NF, Nv, No};
+      int lens[]= {(int)NF, (int)Nv, (int)No};
       if (orbitalPairs){
         lens[0] = NF; lens[1] = Nv; lens[2] = numberOrbitalPairs;
       }
@@ -204,7 +204,7 @@ void FiniteSizeCorrection::calculateRealStructureFactor() {
     }
   }
   else {
-    throw new EXCEPTION("Need Appropriate Coulomb Vertex");    
+    throw new EXCEPTION("Need Appropriate Coulomb Vertex");
   }
 
   Tensor<> *realInfVG(getTensorArgument<>("CoulombKernel"));
@@ -280,7 +280,7 @@ void FiniteSizeCorrection::calculateRealStructureFactor() {
     int aiSliceEnd[]   = {Nv, Nv, orbitalPairEnd,   orbitalPairEnd  };
     int lens[] = {Nv, Nv, numberOrbitalPairs, numberOrbitalPairs};
     realTabij = new Tensor<>(3, lens, Tabij->sym, *Tabij->wrld, "realTabij");
-    (*realTabij) = Tabij->slice(aiSliceStart,aiSliceEnd); 
+    (*realTabij) = Tabij->slice(aiSliceStart,aiSliceEnd);
   }
   else{
     //Get Tabij
@@ -292,7 +292,7 @@ void FiniteSizeCorrection::calculateRealStructureFactor() {
       (*realTabij)["abij"] += (*realTai)["ai"] * (*realTai)["bj"];
     }
   }
-  
+
  //Construct SG
   NG = GammaGai->lens[0];
   CTF::Vector<> *realSG(new CTF::Vector<>(NG, *GammaGai->wrld, "realSG"));
@@ -386,13 +386,17 @@ void FiniteSizeCorrection::calculateComplexStructureFactor() {
     Tensor<complex> *UGF(
       getTensorArgument<complex>("CoulombVertexSingularVectors")
     );
-    int lens[]= {UGF->lens[0], GammaFqr->lens[1], GammaFqr->lens[2]};
+    int lens[]= {(int)UGF->lens[0],
+                 (int)GammaFqr->lens[1],
+                 (int)GammaFqr->lens[2]};
     GammaGqr = new Tensor<complex>(
       3, lens, GammaFqr->sym, *GammaFqr->wrld, "GammaGqr"
     );
     (*GammaGqr)["Gqr"] = (*GammaFqr)["Fqr"] * (*UGF)["GF"];
   } else {
-    int lens[]= {GammaFqr->lens[0], GammaFqr->lens[1], GammaFqr->lens[2]};
+    int lens[]= {(int)GammaFqr->lens[0],
+                 (int)GammaFqr->lens[1],
+                 (int)GammaFqr->lens[2]};
     GammaGqr = new Tensor<complex>(
       3, lens, GammaFqr->sym, *GammaFqr->wrld, "GammaGqr"
     );
@@ -974,7 +978,7 @@ void FiniteSizeCorrection::calculateFiniteSizeCorrection() {
   for (unsigned int i(0); i < GLengths.size(); ++i){
     LOG(2, "StructureFactor") << GLengths[i] << " " << averageSGs[i] << std::endl;
   }
-  
+
   int kpoints(getIntegerArgument("kpoints",1));
   double volume(getRealArgument("volume"));
   double constantFactor(getRealArgument("constantFactor"));
@@ -1022,5 +1026,3 @@ void FiniteSizeCorrection::dryCalculateFiniteSizeCorrection() {
 
   setRealArgument("CorrectedEnergy", 0.0);
 }
-
-
